123
123

+++ Assets/Scripts/EnemyData.cs
@@ -2,5 +2,5 @@ class EnemyData
 {
     public float SpawnTime;
     public float EnemyKind;
-    public float Spawners;
+    public int Spawners;
 }
